requirements.txt
dateparser==0.6.0
humanize==0.5.1
Jinja2==2.10
MarkupSafe==1.0
maya==0.3.3
pendulum==1.3.2
python-dateutil==2.6.1
pytz==2017.3
pytzdata==2017.3.1
PyYAML==3.12
regex==2017.12.12
ruamel.yaml==0.15.35
six==1.11.0
tzlocal==1.5.1
